Fixed a bug with diff tests and added the ability to ignore whitespaces.
The program was hanging if the program was creating too much output.
If the buffer overflows, the program hangs waiting for output to be read.
It has been changed to read output as it is being created allowing the
program to handle output files larger than the size of the buffer.